Revert "added core types; not yet all"
This reverts commit 1c3719472d242e99646f2d4d08ab96efeceb2362.
