+++ module/move/gspread/src/gcore/methods.rs
@@ -5,5 +5,4 @@ crate::mod_interface!
 {
   layer sheet;
   layer values;
-  layer batch_update;
 }
